+++ src/mdlinproto/mdlinproto.go
@@ -90,7 +90,7 @@ type CommitShort struct {
 	Instance int32
 	Count    int32
 	Ballot   int32
-	Status uint8
+	Status int32
 }
 
 // Message types for MultiShard MDL
